services/account-service/docker-requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic[email]
python-jose[cryptography]>=3.3.0
passlib
requests